requirements.txt
flask
flask-restful
flask-sqlalchemy
flask-jwt
pylint-flask
plaid-python